move to folder apps/plugin-browser. migrate (mostly) to mui
