--- apps/hash-browser-extension/src/pages/popup/popup.scss
@@ -1,83 +1,0 @@
-body {
-  margin: 0;
-}
-
-.section {
-  padding: 16px 18px;
-
-  h3 {
-    font-size: 12px;
-    text-transform: uppercase;
-    font-weight: 600;
-    letter-spacing: 0.48px;
-    margin-top: 0;
-    margin-bottom: 10px;
-  }
-
-  a {
-    color: #91a5ba;
-    font-size: 14px;
-    font-weight: 500;
-
-    &:hover {
-      color: #0e1114;
-
-      @media (prefers-color-scheme: dark) {
-        color: white;
-      }
-    }
-  }
-}
-
-.header {
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  background: white;
-  padding-top: 8px;
-  padding-bottom: 8px;
-
-  @media (prefers-color-scheme: dark) {
-    background: rgba(14, 17, 20, 1);
-  }
-}
-
-.avatar {
-  background: #0775e3;
-  border-radius: 50%;
-  color: white;
-  font-size: 18px;
-  font-weight: 500;
-  padding: 8px;
-}
-
-.action {
-  background: #f7fafc;
-
-  button {
-    margin-top: 12px;
-  }
-
-  &:not(:last-child) {
-    border-bottom: 1px solid #dde7f0;
-  }
-
-  @media (prefers-color-scheme: dark) {
-    background: #1f2022;
-    color: white;
-
-    &:not(:last-child) {
-      border-bottom: 1px solid #4d5c6c;
-    }
-  }
-}
-
-.action-title {
-  display: flex;
-  justify-content: space-between;
-}
-
-.options-link {
-  font-size: 14px;
-  text-align: right;
-}
