+++ apps/plugin-browser/.babelrc
