+++ apps/plugin-browser/.eslintrc.cjs
