+++ apps/plugin-browser/LICENSE.md
