+++ apps/plugin-browser/src/assets/img/icon-256.png
