+++ apps/plugin-browser/src/manifest.json
