+++ apps/plugin-browser/src/pages/options.html
