+++ apps/plugin-browser/src/pages/options/options.scss
