+++ apps/plugin-browser/src/pages/popup/popup-contents/sign-in/hash-rainbow-lockup.tsx
