+++ apps/plugin-browser/src/pages/shared/hash-wordmark.tsx
