+++ apps/plugin-browser/src/scripts/background.ts
