+++ apps/plugin-browser/src/scripts/content.ts
