+++ apps/plugin-browser/theme-override.d.ts
