+++ apps/plugin-browser/tsconfig.json
