+++ apps/plugin-browser/utils/env.js
