+++ apps/plugin-browser/webpack.config.js
